requirements.txt
python-telegram-bot==21.6
gspread==6.1.2
google-auth==2.35.0
todoist-api-python==3.2.1
anthropic==0.36.2
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
